Pin third party tools version
At the moment, edb-ansible is not yet ready for Ansible 3, so, we
ensure the installation of the last 2.10.

Azure CLI and Ansible depend on the cryptography module, but the
last version of this python module requires rust for building it.
This is not supported by old pip version, then we pin cyptography
version to 3.3.2, which does not require rust.