inventory.yml and config.yml generation refactor.
Generation of the Ansible inventory file and TPAexec config.yml
can now be delegated to the Python code, through Jinja2 templates.

To make it possible, Terraform must generate a new file: servers.yml.
This YAML file must contains the list of the provisionned servers
with their public and private IP addresses.

Migrating to this new infrastructure is not mandatary: the Terraform
code in charge of DBaaS deployments still generates the inventory.yml
file. This is necessary because in a few cases, the inventory file is
a bit specific.