Follow up fixes
Properly handle applying default values to the specification when
provisioning.  This was only breaking the case when not using a custom
spec file.

Also, configure Vagrant to not replace the default insecure passwordless
ssh keys so we can continue to connect to the VM's without passwords.

Add "dbt2: true" to the inventory file when specified in the spec file.