Updates to virtualbox support (#325)
* Use infrastructure instead of cloud

This is more accurate when describing the non-cloud specifications.

* Add specs command to virtualbox

* Fix virtualbox configure -s option

The --force command was missing, which is required with the -s command.

Don't attempt to merge Null spec values.  Otherwise the validator
breaks.

* Update VirtualBox.md

Remove CentOS7 as it is not actually supported by our tool for the
moment.

* Generate Vagrantfile

In anticipation of being able to customize the resources provisioned by
a spec file, we need to be able to generate the Vagrantfile on the fly.

* VirtualBox: Add DBT-2 support

* Add client and driver information to the spec file
* Process driver and client information from the spec file
  * Save a copy of the spec file in the project directory, required for
    future processing
  * Generate client and driver entries in the Vagrantfile
  * Generate inventory lines for Ansible

* Follow up fixes

Properly handle applying default values to the specification when
provisioning.  This was only breaking the case when not using a custom
spec file.

Also, configure Vagrant to not replace the default insecure passwordless
ssh keys so we can continue to connect to the VM's without passwords.

Add "dbt2: true" to the inventory file when specified in the spec file.

* Update user_spec logic

With the spec file being defaulted correctly, we are supposed to be able
to simply check if user_spec['dbt2_client']['count'] exists.